+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 ### Bug fixes
 
+- [#200][GL-200] - Exception when loading JSON item stack attribute modifiers without UUID
 - [#199][GL-199] - NPE when breaking leaves without any item in hand
 - [#198][GL-198] - Minimum and maximum options for JSON item stacks don't work on Paper servers for Minecraft 1.21+
 - [#171][GL-171] - Some selected scenarios in the scenario menu are no longer displayed with a stack size of 2
 
+[GL-200]: https://gitlab.com/uhccore/uhccore/-/issues/200
 [GL-199]: https://gitlab.com/uhccore/uhccore/-/issues/199
 [GL-198]: https://gitlab.com/uhccore/uhccore/-/issues/198
 [GL-171]: https://gitlab.com/uhccore/uhccore/-/issues/171